// src/services/rest-api/customer-service/model/politic-relation-info-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.267
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface PoliticRelationInfoModel
 */
export interface PoliticRelationInfoModel {
    /**
     * 
     * @type {boolean}
     * @memberof PoliticRelationInfoModel
     */
    'politicianRelation'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof PoliticRelationInfoModel
     */
    'politicianFamily'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PoliticRelationInfoModel
     */
    'politicianPosition'?: string | null;
}